added Orders Services

+++ Entity/Finances/RequestPay.cs
@@ -6,12 +6,12 @@ namespace ECommersShop.Entity.Finances
     public class RequestPay: BaseEntity<Guid>
     {
         public virtual User User { get; set; }
-        public long UserId { get; set; }
+        public int UserId { get; set; }
         public int Amount { get; set; }
         public bool IsPay { get; set; }
         public DateTime? PayDate { get; set; }
         public string Authority { get; set; }
-        public long RefId { get; set; } = 0;
+        public int RefId { get; set; } = 0;
         public virtual ICollection<Order> Orders { get; set; }
     }
